// application/app/api/TMDB/_types/SerieAPI.tsx
import {ContentRating} from "./ContentRating";
import {CreatedBy} from "./CreatedBy";
import {CreditsSerie} from "./CreditsSerie";
import {GenreAPI} from "./GenreAPI";
import {LastEpisodeToAir} from "./LastEpisodeToAir";
import {Network} from "./Network";
import {NextEpisodeToAir} from "./NextEpisodeToAir";
import {ProductionCompany} from "./ProductionCompany";
import {ProductionCountry} from "./ProductionCountry";
import {SeasonSerie} from "./SeasonSerie";
import {SpokenLanguage} from "./SpokenLanguage";

export type SerieAPI = {
    adult: boolean;
    backdrop_path: string;
    created_by: Array<CreatedBy>;
    episode_run_time: Array<number>;
    first_air_date: string;
    genres: Array<GenreAPI>;
    homepage: string;
    id: number;
    in_production: boolean;
    languages: Array<string>;
    last_air_date: string;
    last_episode_to_air: LastEpisodeToAir;
    name: string;
    next_episode_to_air: NextEpisodeToAir;
    networks: Array<Network>;
    number_of_episodes: number;
    number_of_seasons: number;
    origin_country: Array<string>;
    original_language: string;
    original_name: string;
    overview: string;
    popularity: number;
    poster_path: string;
    production_companies: Array<ProductionCompany>;
    production_countries: Array<ProductionCountry>;
    seasons: Array<SeasonSerie>;
    spoken_languages: Array<SpokenLanguage>;
    status: string;
    tagline: string;
    type: string;
    vote_average: number;
    vote_count: number;
    content_ratings: { results: Array<ContentRating> };
    aggregate_credits: CreditsSerie;
}